Add workflow to add the version git tag
This GitHub Action workflow can be called from our application
repositories to add version tags to merge commit. This workflow tags
each merge commit in sequential order with the format `v<number>`.
SemVer compatible format was not used as brings no practical benefit
yet, however versioning format can always be updated in future.

The workflow assumes a pre-existing version tag in which it calculates
the next version number from. The workflow calculates the version number
as last version tag number + the number of merge commit. This allows the
worklfow to be idepotent and independent of workflow run order (which can
be an issue for GitHub Actions).