Fix target git ref in Releases workflow
This fixes an issue whereby the same commit can be tagged with different
versions. The Release workflow is triggered by a workflow_run from a
successful CI workflow. However, git reference associated with
release run isn't the same ref targeted by the CI. By default the ref is
just the branch name, so the Release workflow was reference the HEAD of
the main branch. Therefore, it will try to tag the latest commit rather
than the commit from the merge commit (which CI was targetting).

This fixes updates the checkout ref and target commit for the Release
with an explict reference to the commit sha of the CI workflow run.