Build wget2 binary and add to toolbox image
The wget2 version available for apt package manager is version 1.99,
which contains a bug causing a seg fault when it parses HTML with utf-8
encoding. This builds the latest released version of wget2 version 2.01
and manually adds it to the toolbox image with its library dependencies.

Previously tried to compile a static binary, but a was complex due to
the many dependencies and the lack of pre-built static libraries
available. Also there are dependencies on glibc, which would require
using musl as the standard c library instead.