Build a container image for MongoDB 2.4.
Uh oh. How did we end up here? #poorlifechoices

This is just to give us the option to run our antique version of MongoDB
on Kubernetes temporarily until we finish the port of Router API to
Postgres. It's helpful to have that flexibility because switching off
the old Puppet stack quickly is more important than getting Router API
on to Postgres quickly.

The idea would be to run a StatefulSet of these and do a quick
stop-and-swap while nobody's publishing anything. ([Ah, the old
stop-and-swap](https://github.com/alphagov/govuk-infrastructure/assets/1170635/840cdc5f-49e5-4b30-a377-c6941e5a0b57).)

This thing isn't in the serving path and untrusted input doesn't go
anywhere near it, otherwise we wouldn't even be considering this gross
workaround. If we do it, it's going to have to come with a hard deadline
so that we're not left holding a bag of very stale MongoDB.

Hopefully we can just get the Router API -> Postgres port back on track
and not need this, but it's good to have options. Perhaps manifesting
the horror of this putative workaround will galvanise the porting
effort? We really don't want to find ourselves running this image in
2024. Double plus ungood.

Differences from the [last docker-library/mongo image
release](https://github.com/docker-library/mongo/tree/4507e97/2.4):

- image manifest format that works with containerd (so it runs on k8s
  ≥1.24)
- up-to-date base image, libstdc++, etc. etc.
- we don't care about NUMA (our cluster nodes aren't)
- ditched the entrypoint shell script (we don't need it)
- runs as non-root by default

Tested: Router end-to-end tests pass with this image.