Add opt-out of repo permission granting
Adds a mechanism for opting out of the access-granting, e.g. a GitHub topic govuk-sensitive-access
would indicate that the repo needs access to be managed manually, and so not to try granting access
to the teams. We anticipate using such a mechanism for github-action-dependabot-auto-merge,
where access to these teams should not be granted for security reasons.

https://trello.com/c/tbQjgkwR/3249-automate-granting-team-access-to-govuk-tagged-repos-3