Remove unnecessary team overrides for Sentry projects
Historically, teams have only really been used to coordinate
alerts for team-owned projects, into their team Slack channels.

In all of the teams removed in this commit, no alerts were set up
for the projects in question, so there is no value in keeping them
as named 'owners' of the apps in Sentry, beyond the access they
already have by default in the '#govuk-developers' team.