Fix db-backup IAM role policy.
This is the IAM policy for the new environment sync / backup jobs that
will replace govuk_env_sync[1].

The permissions are similar to those for govuk_env_sync, but tightened
up a bit so that a user with integration admin has no access to
information that's only supposed to be in production/staging.

- Each environment (production, staging, integration) writes to its own
  S3 bucket only.
- Production can read from both of its buckets (the main one and the
  out-of-region replica, so that we can restore from there if we have
  to).
- Staging can read production backups as well as its own.
- Integration can read staging backups (which should contain no PII or
  unpublished content) as well as its own.

[1]: https://github.com/alphagov/govuk-puppet/tree/main/modules/govuk_env_sync

Tested: applied in staging.