Document reusable actions consistently.
It didn't make sense to have a usage example for just one of our many
reusable GitHub Actions in the top-level README, while not mentioning
that we even have a library of reusable actions in here.

Link to all the reusable actions/workflows from the top-level README
rather than only mentioning one of them.