Fix IAM policy for aws-load-balancer-controller.
This had gotten out of sync with the version of aws-lb-controller that
we're running.

This time, just take Amazon's policy file verbatim instead of formatting
it as HCL. This way it'll be easier to diff next time.

The `jsonencode(jsondecode())` is not a no-op; it gives us plan-time
syntax checking of the json policy file.