Make CloudFront include Host header in cache key
Our CloudFront failover distributions have publicly accessible domain names (e.g. `xyz.cloudfront.net`). If a user visits any of these URLs, the request is forwarded to origin, which returns an HTTP 404 (because the host is unrecognised), which CloudFront caches.

Prior to this PR, future requests made to the same path but with the correct `Host` header (e.g. `www.gov.uk`) would then serve the cached HTTP 404, instead of the correct response.

This PR configures CloudFront to include the value of the `Host` header in the cache key, so that requests made to the distribution with different `Host` headers are considered different requests, and are cached separately.