Change size of shared Redis in staging
We are currently using `cache.t4g.medium` for the shared Redis in
staging. This provides 3.09GB of memory.

Over the last weekend, we saw instances of the staging Redis run out of
memory when batch jobs were running.

Therefore changing the size to `cache.m6g.large`, which matches
integration and gives 6.38GB of memory.