Workflow users can provide additional image tags
It takes a space-separated string.

This change is motivated by a desire to deploy separate images to each
environment but for a non-app project (i.e. doesn't use Argo workflows
or `image-tags`).

We only need a single additional image tag on our specific project, but
to me, the future proofness of this option felt like it outweighed the
additional complexity.

(I've tested this with an empty value, a single tag and multiple tags,
and it's shown up with/without them as expected in ECR for all of them.)