Add Sentry project for `govuk-sli-collector`
I wasn't aware that this existed and I'm still not sure what it's going
to do. But I noticed that other projects are here and assume that this
should be, too.

I've actually already created the project in Sentry*. After seeing this,
I renamed it to start with "app-", though I don't know if that'll help
or hinder.

*https://govuk.sentry.io/settings/projects/app-govuk-sli-collector/