Remove a couple of not-very-useful log fields.
Arch is always amd64 for the time being (and when we get around to
arm64, we can put it back if we actually end up wanting to run mixed
clusters during the transition, which is unlikely).

Instance is not very useful, because it'll only differ from
`kubernetes.labels.app_kubernetes_io/name` or
`kubernetes.deployment.name` if it's a manual test install of a chart,
in which case the whole name will be different anyway.

This should help a bit with log quota.