Add Clamav image that can run unprivileged
The current official Clamav image cannot be run unprivileged by default.
This adds a script and changes the file permissions, so that it can run
as non-root. At time of writing, there is a pending PR to add this
functionality to the offical image:
https://github.com/Cisco-Talos/clamav-docker/pull/33. Once this is
available we can switch to the offical image.