Log Snyk findings to stdout
Previously we were saving the results to a JSON file and archiving this as a build artefact.

Snyk returns error code 0 if the scan was successful but found no issues. It returns error code 1 if the scan *was successful* but issues were found. This combined with the lack of log output is surprising, and makes it seem like the build failed when it actually succeeded.

As we're not currently using the JSON output anywhere, we can just stop generating it and log the results instead.