Make GitHub CI user token available to Dependabot
As of version 4 of `delete-artifact`
(https://github.com/GeekyEggo/delete-artifact), a GitHub API token is
required to remove artifacts, which we are doing as part of the CI
pipeline.

This secret is included for access by GitHub Actions triggered events,
but is not available to Dependabot.

Therefore this step fails on PRs raised by dependabot.

This makes the secret accessible to dependabot.