Enable caching for image builds
This replaces our custom implementation to build and push images with
Docker's own GitHub Action. From this we gain ability to cache layers
using GitHub's cache and to decrease workflow runtime. This change
removes the "additionalBuildArgs" input, which was only being used by
licensify. This is replaces with "buildArgs" which is more specific in
its usage.