Fix image tags generation for non released versions
When trying to deploy an gitRef that wasn't a release version (i.e. not
`v*`), for example a branch deploy, the sha used as the image tag
incorrectly referenced the HEAD commit on main. This was because the
docker/metadata action generates the sha tag using the `github.sha`
workflow context. For workflow dispatch events this is the HEAD of main.

This fix now finds the SHA of the local head for the sha image tag.