Rename "CSVs" to "reports" to better reflect actual usage.
Several "publishing" (i.e. CMS) components write (non-sensitive) reports
to S3 buckets. Some of these happen to be in CSV format.

Currently, the permissions for writing to these report buckets are split
arbitrarily (for purely historical reasons) between a handful of
different IAM policies all attached to the node role.

It makes sense to consolidate these for now, at least so that it's
easier to see where to add/change stuff (for example recently in
dea2db5).