Elide an unnecessarily separate IAM policy for pubapi.
There's no benefit in having a separate, almost-identical IAM policy
for Publishing API to write event logs to S3 when both policies are
attached to the node role anyway.

If we were to someday split up the roles, assign them to serviceaccounts
and use pod identity (almost certainly overkill in this case) then it
would make sense to have separate policies. Until then, it's just
additional toil and potential for confusion (which itself is not good
for security).