Make it easier to keep up with Terraform releases.
We want to be on the latest patch release by default, i.e. unless we're
specifically overriding it.

We want to be able to update minor versions with minimal fuss, by
changing the version in one place and having it affect all workspaces
that aren't overriding the version constraint.

We want sufficient flexibility in holding back major version updates,
on individual workspaces/projects where necessary, without introducing
toil.

I think passing a terraform_version variable into each instantiation of
workspacer achieves this, because we then have a global default that we
can override per workspace if necessary.