Make Snyk SCA scan failures block builds
This is a requirement before we can implement RFC-167, "Allow automatic
patching of all dependencies".

If this change causes problems on any our repos, we could temporarily
patch those repos (and those repos only) to refer to the previous version
of this workflow instead of `@main`. For example:

```yaml
jobs:
  snyk-security:
    name: SNYK security analysis
    uses: alphagov/govuk-infrastructure/.github/workflows/snyk-security.yml@533143ddd823518b97165a4a80c2995e9ffe980a
    secrets: inherit
```