Read GH token from secret in autorelease workflow
We don't allow GitHub Actions workflows to raise or approve pull
requests on any of our repos, so this workflow will need to use a token
with the appropriate permissions instead of an auto-generated one.

We already have an org-level secret with the appropriate permissions,
so let's use that one.