Cache Snyk binaries in snyk-security workflow
Updates the snyk-security workflow to cache the Snyk binary (and the
/usr/local/bin/snyk wrapper script) between runs of the Snyk workflow.

We don't want to cache the Snyk binary for more than say 1 day, otherwise
we'll end up running an old version indefinitely. The actions/cache
action doesn't allow you to specify an expiry time, so we instead include
the current date in the cache key.

We include the runner OS in the cache key because the path to the Snyk
binary varies depending on OS. For example, on Linux, the Snyk binary
lives at /usr/local/bin/snyk-linux.