Turn off Argo CD chart's "redis-ha" option.
This option sounds great, but turns out to be pretty gross under the
hood and is almost certainly a net loss for reliability.

Proximate reason for getting rid of it is that it doesn't play nice with
the AL2023 node image. This bug is just the tip of the iceberg:
https://www.github.com/argoproj/argo-helm/issues/1958

Redis isn't really critical for Argo CD anyway; it's fine if it's
momentarily unavailable, whereas running a bunch of HAProxy in front of
a replicated, leader-elected, Redis statefulset is a recipe for pain.