Create a shared action for running a node script
This creates a new reusable GitHub workflow that can be used to run a
node JS script from a repository. By script I expect it to be something that
is defined inside a package.json scripts object (e.g. yarn run lint:js).

It then utilises this workflow on the standardx and stylelint workflows
to simplify them.

The motivation for adding this was due to on govuk-chat deviating from
standardx to standard and then having to define our own workflow [1]. I
looked and saw other repos had also defined their own near identical
workflows [2] [3] so felt the govuk-chat need was not unique.

A demo of this working from a branch is available at: https://github.com/alphagov/govuk-chat/pull/236

[1]: https://github.com/alphagov/govuk-chat/pull/235/commits/fbb629e60854fc51808abb8867d84679ca30e1f9
[2]: https://github.com/alphagov/whitehall/blob/main/.github/workflows/lintjs.yml
[3]: https://github.com/alphagov/whitehall/blob/main/.github/workflows/lintprettier.yml