Give GitHub Actions access to pact secrets for some repos
For repos tagged `pact-publisher` (currently only `gds-api-adapters`) we
need to give GitHub Actions access to organisation level secrets for
accessing the pact broker, in order to publish pact tests.

This was never an issue before as the pact tests was incorrectly set to
only publish from a PR being merged into the main branch, so were never
run on automatically raised PRs (e.g. dependabot). This was resolved in
https://github.com/alphagov/gds-api-adapters/commit/5d27c95d96fb49334b002c3be70993e6b6f1145e.