Add workflow to enforce gem version bumps
This workflow will run when opening a pull request on gem repositories.
It checks if there are changes to the CHANGELOG.md and version.rb/*.gemspec files.
If those files have not been changed the workflow will fail.

Code changes are often merged into gems without the gem version being bumped.
This leads to changes not being released quickly and sometimes piling up.

https://trello.com/c/R1kkdVjt/3560-re-evaluate-effectiveness-of-gem-auto-release-workflow-3