Explictly list GH repos that can write to ECR.
We were using GitHub Topics (labels that can be applied to a repo) to
determine which repos should be permitted to write to our ECR container
registry.

Turns out this is inadequate, because anyone in the alphagov GitHub org
who can create repos (roughly 700 people, most of whom don't
work on the GOV.UK web CMS) can create a repo and set whatever topics
they like on it, then get write access to our private ECR.

For now, let's just go back to maintaining a list of authorised repos.
That shouldn't be too toily in the short term, since we don't create new
repos all that often.