Split IAM policy for ECR push per GitHub repo.
This generates a nasty quantity of IAM guff, but this is the next
simplest thing I can think of since the 2 kbyte assume-role/trust policy
limit scuppers the obvious solution of just listing all the repos in a
condition.

This way also restricts a GH repo to writing to its own repo in the
container registry, so at least we get something in return for the
clutter.