Ignore datastore `solution_types` property drift
Some properties of some resources on Vertex AI Search can permanently
change based on tangentially related changes to other resources, leading
to Terraform state drift.

For example, we created a recommendation engine on integration for
testing purposes, and it permanently changed the internal state of the
datastore it was attached to (despite the engine since having been
deleted).

This ignores any post-creation changes to the `solution_types` property
to mitigate any consequences of this undesired behaviour (in this case,
Terraform trying to recreate the resource as it cannot return the field
back to its original state).