Enable ghcr.io pull-through cache for ECR
This commit creates a pull-through cache rule to pull images from
ghcr.io. This allows us to push images to ghcr.io instead of directly to
AWS ECR. We need to explicitly create the repositories for the
pull-through cache, as ECR doesn't support cross-account permissions to
create the ECR repositories on the fly.