Unpin content-data-api minor versions
We have auto minor version upgrade set on these databases, but terraform
is trying to pin them to a specific minor version. That means that
currently, terraform applies try to "upgrade" from 13.15 to 13.13, which
AWS won't allow.

All the other postgres databases just specify the major version,
avoiding this issue.