Fix names of kubernetes nodes
Currently in AWS each of the arm nodes is just called "arm", and each of
the non-arm nodes is called "main".

It looks like the eks module is pulling these from the key (main / arm)
in this map.

We switched from using name to name_prefix in fa08608c, but the
[documentation for the eks module](https://github.com/terraform-aws-modules/terraform-aws-eks/tree/master/modules/eks-managed-node-group#input_name)
doesn't mention name_prefix - it only has name and use_name_prefix.

Switching to name means the module will name the instances with
something more descriptive than "main".