Replace variables for chat redis version
This replaces the ability to change Redis version in each environment
and have different parameter groups.

The motivation for this is that we want to set specific parameters to
tune Redis for sidekiq and this will be coupled to the redis version.

This approach is also consistent with the approach for Redis version on
the shared redis instance.