Add API Gateway config for search
This is a REST API because it allows [the following advantages](https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-vs-rest.html) over an HTTP API:
- edge optimization (CloudFront integration)
- WAF
- rate limiting/throttling
- cache control
- request validation/modification

Not all of these are immediately necessary, but they are likely to be used in the near future.

The stage name is "v0.1" which indicates that the API is subject to change. THis results in a URL like example.gov.uk/v0.1/search.json as the stage name is automatically injected into the path.