Add custom domain and WAF for search-api
The WAF needs to be applied in a similar way to CloudFront WAFs which requires the domain configuration.

This applies to the API gateway we're setting up, not to internal traffic.

It permits 100 requests per client per 5 minutes, which is 1 every 3 seconds on average. This is probably sufficient for now, though if lots of users are behind a single IP, then this will need amending.
This WAF also applies across the whole deployment, so if we add another API that is likely to be more frequently used, we'll definitely want to relax it and do more management at the API layer.

This also registers the API Gateway with AWS Shield Advanced.