Auto eval: remove autocomplete evaluation branches
The `collate_results` function takes a `type` parameter which is used to
determine where to source the results from.

Currently the conditional caters for 4 different `type` strings and
calls the relevant service for the results.

This `type` param comes from a list of `candidates` which is
defined in the `config.tftpl` file. Each candidate has a type and the
script loops over the array, ultimately calling this function with the
corresponding `type`.

However the config file only lists a single entry in the candidate
array, with a `type` of `"vertex:search"`. So the other three branches
of the conditional will never be executed.

To make porting this script over to Ruby a bit easier, we can remove all
the code paths that will never be touched.