Allow secrets to be set when building image
This adds an optional configuration for allowing the calling workflow to
set one or more build secrets, which get passed in to the `build-image`
step.

The driver for this is the GOV.UK Chat Rails app which has a private gem
dependency. In order to run `bundle install` when building that app, we
need a secret token set, which we'll access via secret mounts in the
Dockerfile[1]. In order to mount the secret, we need to pass it in using
the `secrets` option in the docker/build-push-action action[2].

This change is backwards-compatible with existing callers of the
workflow, as omitting the secret has no effect on the build step.

[1] https://docs.docker.com/build/ci/github-actions/secrets/
[2] https://github.com/docker/build-push-action?tab=readme-ov-file#inputs