Search v2: Remove superfluous env configuration
This removes all of the `DISCOVERY_ENGINE_*` env vars from the Secrets
Manager secrets for Search API v2 and Search Admin except for the
`DISCOVERY_ENGINE_DEFAULT_COLLECTION_NAME` (which the apps now infer the
other values from).

Note that we're leaving the datastore in the module outputs (and adding
in the default collection) even though the outputs are not used by other
modules, just because it's handy to see those (non-sensitive) values in
the Terraform Cloud UI for a quick sense check.