Add retry and wait_to_retry support for Search API queue
This adds some new configuration to allow messages on the
`search_api_to_be_indexed` queue to be retried by the consumer.

For context, this is to support an effort to remove Sidekiq from Search
API. Currently we consume messages then enqueue them directly on Sidekiq
for processing. This is sometimes causing issues where Redis crashes if
lots of jobs are being processed at once. The aim is to remove Sidekiq
and process all jobs in the consumer directly.

However in doing this, we'll lose the automatic retry functionality that
Sidekiq provides. So instead we'll implement our own retry mechanism
using RabbitMQ only.

The idea is as follows:

* A consumer picks up a job from the `search_api_to_be_indexed` queue
  for processing
* If an exception is raised, it's caught and the message is rejected
* The `search_api_to_be_indexed` queue is now configured to have a new
  dead letter exchange of `search_api_to_be_indexed_retry_dlx` where
  messages have a specific TTL of 60 seconds
* We have a new `search_api_to_be_indexed_wait_to_retry` queue bound to
  that dlx
* Nothing is listening on the `search_api_to_be_indexed_wait_to_retry`
  queue, so messages are never processed. Instead, when their TTL
  expires, they're routes to another dlx,
  `search_api_to_be_indexed_discarded_dlx`. This has a binding back to
  the original `search_api_to_be_indexed` queue which applies to all
  messages.
* Now messages are back on the `search_api_to_be_indexed` queue they're
  picked up by the consumer again. The consumer can then decide whether
  it wants to try to process the message again (e.g. by inspecting the
  retry count)