Search API v2: Remove `move` block and update `imports`
This removes the `move` block for the moved completion config now that
it has successfully been applied across environments.

We originally managed "static" resources that always exist on VAIS by
overriding the REST API provider's `create_*` properties to be the same
as the corresponding `update_*` ones, but a semantically more meaningful
way of dealing with these is through Terraform's `import` blocks. This
adds an `import` block for the completion config.