Search API v2: Add `cdoc_url` to data store schema
This adds a `cdoc_url` field to the data store schema, which we will
populate with the document's `content_id` value.

This field doesn't have to contain a URL, just an arbitrary unique
reference of our choice. We decided that since we already use
`content_id` as an identifier across the board, we will use it for this
purpose too.