Search API v2: Add config for global engine
We've recently added a `govuk_global` engine to our setup, with the
intent of replacing the existing `govuk` engine with it.

This commit duplicates the existing `govuk` engine's serving config
and controls to ensure the new engine is configured in the same way.
It uses the new `fixedBoost` and `dataStore` attributes on controls
to be in line with how controls work now (the existing ones use
legacy attributes).