Search API v2: Avoid resource drift
We have a number of resources that continuously present drift false
positives due to the API returning additional or subtly different
(e.g. `null` instead of `[]`) values compared to what we set in the
resource definitions.

This has been exacerbated by a recent major update of the REST API
Terraform provider, which now refreshes resource state based on
remote state.

This uses a combination of the REST API provider's `ignore_changes`
attribute, more sensible default values, and adding of missing values
to work around some of this drift.