Search API v2: Simplify apply process
This simplifies the way Terraform changes are applied by removing the
superfluous run triggers (which cause double runs to be queued up for
staging and production as one already gets queued on VCS change
anyway).

It also changes the auto-apply logic to always autoapply except in
production, which means that staging no longer needs to be applied
manually going forwards (saving a bit of toil).