Add ithctester K8s group, cluster role and EKS Access Entry
This role grants ITHC testers read-only access to most resources in the cluster. This includes some custom resources, such as ArgoCD Applications, Argo Workflows and external secrets. This does not allow testers to read the contents of secrets.

Access is granted by the AmazonEKSViewPolicy access policy, which grants
read-only access to most K8s resources. Extra permissions are granted
via a cluster role, so testers will be able to access custom resources
that are not referenced in the Amazon-managed access policy.