Suffix publishing infrastructure vars sets with "-non-sensitive"
When the Terraform for these groups ran for the first time we discovered that
groups with exactly the same name existed in the sensitive vars sets defined in
another repository. Those var sets are properly named within that repo, so we
suffix these ones with "-non-sensitive" to avoid the clash.