Prevent the readonly role from getting secret values in the cluster
The readonly IAM role assigned to humans prevents them from reading secret
values, but the equivalent Kubernetes role did not. We use the external secrets
operator, so being able to read inside k8s is equivalent to being able to read
in AWS.

The readonly role is now limited to listing which secrets exist, but not
reading their values, within the cluster.