Allow duplicate DNS record names in publishing infrastructure
Previously we used the record name as the key for the record in Terraform, but
this has the side effect of not allowing duplicate names, which exist when we
have >1 type of record for a name (such as an A record and a TXT record).

To work around that, we include the record type in the key. The combination of
name and type will always be unique.

This will make Terraform destroy and recreate the records, which is OK because
this zone isn't in use yet.