Exit on errors in shutdown.sh
From https://www.shellcheck.net/wiki/SC2181

> Running a command and then checking its exit status $? against 0 is
> redundant.

> Apart from the redundancy, there are other reasons to avoid this
> pattern:
>
> ...
>
> - Scripts that run or are called with set -e aka errexit will exit
>   immediately if the command fails, even though they're followed by a
>   clause that handles failure.

We can also avoid the call to grep just by using bash's builtin support
for regex checks. We need to use `[[` instead of `[` for this, but since
the shebang says `bash` anyway we don't have to worry about
compatibility with cshell or whatever.