Add test path to filtered_pages
There are only two remaining pages being filtered from search results. If a request is made to remove them, we are unclear how an empty array will be parsed by terraform and so would at that point remove the whole block.

Removing the block will make it less easy for non-team members to implement filters in the future. So the most pragmatic thing to do is probably to ensure that a placeholder is always present.

This code is a prime candidate for tooling in search admin.