Create bucket for automated evaluation output
Trello: https://trello.com/c/byVjO4ZU/876-surface-vais-evaluation-results-in-bq

The best name is already in use for the "old" evaluations. This is for the out-of-the-box evaluations now provided by Vertex AI, hence naming it `vais_evaluation_output` for "Vertex AI Search".

Steps to follow once data is available in the bucket:

1. Create a BigQuery dataset similar to https://github.com/alphagov/govuk-infrastructure/blob/eccb6203f57bf3b68e5fe2d601360ef1c4269ad4/terraform/deployments/search-api-v2/automated_evaluation.tf#L164
2. Create one or more BigQuery tables similar to https://github.com/alphagov/govuk-infrastructure/blob/eccb6203f57bf3b68e5fe2d601360ef1c4269ad4/terraform/deployments/search-api-v2/automated_evaluation.tf#L237