Add govuk_chat_private as a repo
This adds the repo checks. The regular standard security checks CodeQL
doesn't work for a private repo and the dependency-review one seems to
require additional set-up beyond what a GitHub admin can do. So both are
disabled.

The terraform import statement is added to prevent terraform trying to
create the resource - it was a suggestion from the reviewer, I don't
understand it.

The govuk_chat_private gem isn't a conventional gem as it is not pushed
to RubyGems but it does need the `GOVUK_CI_GITHUB_API_TOKEN` so I've
configured it so [1].

[1]: https://github.com/alphagov/govuk-infrastructure/blob/a19e6c85656987ad450f7da559f0103a1eecbe3a/terraform/deployments/github/main.tf#L255-L258